Fixes for FreeBSD deprovision
When removing the previously provisioned user the 'rmuser' command wants
to kill all processes that are owned by the user, but this kills the
user's shell and 'waagent -deprovision+user' command before we're able
to properly deprovision.  We fix this by moving the DeleteAccount()
routing a bit lower in Deprovision(), and run rmuser in a detached
process to prevent it from killing our python script.