Fix issues with resource disk mount (#600) (#629)
* Check mount output to detect already mounted resource disk (#600)

mount(8) returns '32' in a variety of cases, we need to check the output
to be sure the resource disk is already mounted.

* Pass '-n' to sfdisk when not writing

Some old versions of sfdisk(8) will open block device in O_RDWR mode regardless
of the face that we're actually just querying partition type and not changing
it and if device is opened in this mode it will trigger partition rescan when
it is closed. This creates a race as partitions (/dev/sd*) will be going away
and re-appear later. Specify '-n' to indicate no write is required, sfdisk(8)
always opens the device in O_RDONLY now.

* Re-read partition table after changing partition type

Some old sfdisk(8) versions don't do syncronious partition re-read after
making changes. The kernel will re-read it anyway in background but this
will cause races with further mkfs as partition devices will be
disappearing. Do explicit partition rescan after we change partition id.