Fix clearlinux distro_name (#792)
Starting from the release 16180 Clear linux changed the NAME in
/usr/lib/os-release from "Clear Linux Software for intel Architecture"
to "Clear Linux OS for Intel Architecture", with this change the
waagent service is crashing because it doesn't find the waagent.conf
file due it doesn't detect the Clear Linux distroname.

This commit adds support for previous and current Clear Linux OS for
Intel Architecure releases by searching the "Clearl Linux" string
in osinfo[0] and returning clearlinux in distro_name.

Fixes #792

Signed-off-by: Ornelas Aguayo, Jesus <jesus.ornelas.aguayo@intel.com>