Fix brokenness caused by building an extension in Windows
If the extension zip was generated on Windows, it may not have the
proper Unix permission bits set. It may have weird newlines which need
to be replaced.

Perform python equivalent of chmod a+x. Reading and writing the file to
a python string is sufficient to fix the weird newline characters.