bump version count from 5 to 20
Some of the tests hard code expectations about the agent version when doing
upgrade tests.  This hard coded value was set at a particular version of the
agent. We have since upgraded past that version, and now tests are breaking.
The test created five versions, so the simple fix is to increase the value to a
large number.  A better fix is still required, and is planned.