Create scripts/tests.sh and add scripts/pre-commit symlink
Travis uses this tests script now. I added a symlink for the pre-commit
hook but this could be changed to a script which does stuff other than
run the unit tests before creating the commit.

After cloning the repo, one must symlink .git/hooks/pre-commit to the
pre-commit script to activate it.