Handle set_handler_status failing
A object constructed in memory fails to be converted to a JSON string
representation.  This is causing failures in a minimal number of cases.  The
error is now handled, and logged. A stack trace is now captured if an exception
is thrown to better improve debugabilitiy.

Further define the behavior of file_write with a test case the demonstrates how
to behave when contents is None.