Add support for provisioning SRIOV Infiniband VMs (#1365)
* Introduce nd_version in RDMA handlers

To prepare for adding support for SRIOV based Infiniband, add nd_version in RDMA handlers. nd_version is passed from the KVP for Network Direct based Infiniband VMs. This is used as an indicator to decide if we should go provisioning Network Direct Infiniband, or provisioning SRIOV Infiniband.

* Introduce install_driver_if_needed to install drivers only necessary

Unlike Network Direct Infiniband VMs, we don't need to install any drivers for SRIOV Infiniband VMs. The VF for SRIOV Infiniband is already built into the kernel. Change the code path to install drivers only if Network Direct environment is detected.

* Add support for provisioning SRIOV Infiniband

Unlike Network Direct Infiniband that exposes an ethernet NIC for communicating with the host on routing and address resolution, SRIOV uses IPoIB for the same purposes. Add support to setup IB address as soon as the device shows up in Infiniband driver sysfs, and IPoIB stack is ready.