Add helper function enable_check_rdma_driver
We already have OS.CheckRdmaDriver defined for waagent config, add a helper
function to read its value.

The default returned value is True if OS.CheckRdmaDriver is not defined in
waagent.conf, this is for keeping the behavior of older versions of waagnet
unchanged.

On newer HPC images, checking and updating RDMA driver is not necessary. The
user can elect to set OS.CheckRdmaDriver=n to skip those steps.