Removed threading for RDMA
There are issues where about 5% of nodes fail to initialise the RDMA on a reboot.  The problem shows as a defunct process that was created with the run_get_output function in the shellutil.py.  This is when Python 2.7 is used and is probably related to the thread safety issues of the subprocess module.  This stackoverflow page gives a description and links to bugs - https://stackoverflow.com/questions/21194380/is-subprocess-popen-not-thread-safe.  The thread seems an unnecessary optimisation as the RDMA initialisation runs in under 0.1s.

Note: this should not be an issue with Python 3 but this is not standard for most OSs.  Another alternative could be to make it Python 3 only and release the azurelinuxagent as a packaged executable.