Correct the status reporting to be based on extensions's status'
exthandler.py - Now the extension errors would only be WARNINGS and not ERRORS. Also, if the extension status is too big, we still parse it to get the actual status of the extension and make it leaner. Also included use of NamedSet.
exception.py - Added a new exception type
util.py - Updated NamedSet to be used in other places.