Tentative fix for Bug 7361722 (Check ExtHandler state before running disable command).
During extension version update, the windows GuestAgent only envokes
the extension "DisableCommand" when the extension is in the "Enabled"
state. This commit's purpose is to mirror this functionality in the
linux GuestAgent.
"Tentative" because no testing has been done yet; the code must be
checked into a repo before it can be cloned onto a VM (for live
testing).