Isolated file-specific fetching logic from the WireProtocolData class.
Renamed the WireProtocolData class to WireProtocolDataBase, and created two
derivative classes: WireProtocolDataFromFile; and WireProtocolDataFromMemory.
The majority of functionality remains, shared, in the base class. The original
reload function was moved to the WireProtocolDataFromFile class, and a new one
written for its counterpart. The only difference between these two functions are
the way in which they access data. Additionally, added an optional parameter to
the mock_wire_protocol function in mocks.py that is used as a factory method to
select the WireProtocolData* derived class coinciding with the format of the
inputted data.

Separated the filename dicts into their own file, and created another file to hold
a basic implementation of the "data fetcher" used by the new dynamic data class
(mockwiredata_fetcher.py).