Simplified changes to Unit Test Framework (#1969)
* Tentative fix for Bug 7361722 (Check ExtHandler state before running
disable command).

During extension version update, the windows GuestAgent only envokes
the extension "DisableCommand" when the extension is in the "Enabled"
state. This commit's purpose is to mirror this functionality in the
linux GuestAgent.
"Tentative" because no testing has been done yet; the code must be
checked into a repo before it can be cloned onto a VM (for live
testing).

* Added Unit Tests confirmed correct behaviour for version updates.

Two tests were added; the first makes sure that, like the prior behaviour,
an enabled extension is disabled when performing a version updated; the second
checks that an extension in a non-enabled state (e.g. Installed) is not disabled
upon version update.

* Isolated file-specific fetching logic from the WireProtocolData class.

Renamed the WireProtocolData class to WireProtocolDataBase, and created two
derivative classes: WireProtocolDataFromFile; and WireProtocolDataFromMemory.
The majority of functionality remains, shared, in the base class. The original
reload function was moved to the WireProtocolDataFromFile class, and a new one
written for its counterpart. The only difference between these two functions are
the way in which they access data. Additionally, added an optional parameter to
the mock_wire_protocol function in mocks.py that is used as a factory method to
select the WireProtocolData* derived class coinciding with the format of the
inputted data.

Separated the filename dicts into their own file, and created another file to hold
a basic implementation of the "data fetcher" used by the new dynamic data class
(mockwiredata_fetcher.py).

* Added ExtensionActor and friends to help make dynamic mockwiredata generation
more concise.

Baked in support for calling mock_data_protocol() and patch("subprocess.Popen"),
as these seem to be the current "accepted" way to set up tests.

* Fixed VersionUpdateTestCase tests to use the new framework.

* Added todo stubs for code cleanup.

Most todo stubs are for comments and docstrings, though a couple also
request code changes.

* Added comments explaining new code.

Still remaining are some TODO's dealing with code changes in
extension_actor.py.

* Added note pertaining to parameter limitations for ExtensionActor
factory.

* Cleaned up tests.

* Reverted class name change that broke a lot of tests.

class WireProtocolData was changed to class
WireProtocolDataFromFile, which broke every test that
directly referenced that class.

* Finished various TODOs.

* Reverted test_exthandlers.py, which need not be updated.

* Updated docstring for mock_wire_protocol function, documenting the
possible parameter configurations.

* Fixed Initial syntax errors for python versions <3.6 (2.6-3.4, specifically.)

* Replaced dict comprehension with python2.6 compatable for loop.

* Replaced 'with ZipFile(...)' with open(); try; finally: close() paradigm,
to be compatible with 2.6.

* Updated assert commands to self.assert{Equal, NotEqual, True} commands.

* Fixed most variable names to conform to PEP, merged ActionSet
and ExtensionInfo classes.

* Changed TextExtensionUpdateOnFailure._do_upgrade_scenario to raise
instead of assert for semantic reasons.

* Updated more variable names to conform to PEP.

* Fixed imports in middle of function.

* Added example to replace_xml_attribute_value_on_matching_element
and attempted to match the regex groups to said example.

* Updated filename manipulations to use os.path.* functions where
appropriate.

* Finished 'TODO: Desc's and fixed potential bug with patched_popen's
getting stopped in the wrong order.

* Finished last TODO: Desc.

* Added comment explaining why the mock_http_get doesn't need to
check the first actor.

* Fixed typo and replaced "import *" with explicit import.

* Added support for guaranteed invocation ordering.

* Removed redundant test.

* Adding initial simplified features.

* Removed test added automatically by merge.

* Replaced zipfile mocking with handlermanifest mocking.

* Added python2* compatibility.

* Finished up code.

* Added check.

* Added enum backports to test-requirements.txt.

* Fixed empty exception text.

* Fixed comments.

* Removed dependency on enum34 (python ver < 3.4)

* Fixed problems mentioned in PR comments.

* Removed enum import

* Fixed final issues.

* Fixed python2 compatibility issues

* added functionality to dynamically determine the written extension status.

* Fixed duplicate arg names error.

* Suppressed error we don't want to listen to the linter on

* Fixed pylint errors.

* Fixed python2 compat. issues (filter doesn't always return an interable in 2.x)

Co-authored-by: Ubuntu <kevinclark19a@kclark-ubuntu1804.k3mre2kg3euulf11q5ukpfplwe.jx.internal.cloudapp.net>
Co-authored-by: keclar <keclar@microsoft.com>