Add LIS version to OSInfo.message (#2011)
* Add LIS version to OSInfo.message

This uses the Linux kernel’s `modinfo` command to retrieve the “version”
field for the “hv_vmbus” kernel module (the LIS drivers). This is the
documented method to retrieve the LIS version for the module.

This command should not hang or be missing, given that it’s part of the
Linux kernel. Moreover, the existing RDMA code already relies on
`modinfo` without a timeout etc. It is also distribution-agnostic.

If the system does not have “hv_vmbus” as a module, or the command fails
for some other reason, we return “Absent.” In the case of an exception,
we return “Failed.” Note that for systems where the driver is built into
the kernel, or installed as a module that the LIS developers did not
create, it will report “Absent” as it depends specifically on our
provided module’s “version” field. In the former case the driver version
can be deduced (and is tied to) the kernel version, and in the latter
case it can be deduced from the distribution’s package.

While a timeout would be nice, the existing `run_command` doesn’t
currently support it (and adding it will require some work because we
can’t use Python 3.3’s `timeout` due to Python 2.6 being our minimum
version). Again, there is already existing code which calls `modinfo`
without a timeout, so if this were to fail, it seems it would fail
regardless of this patch.

* Fix majority of disabled pylint errors in test_version.py

It did not make sense to fix one new test, which would then be
dissimilar from all the other tests, when these could be fixed with a
few minutes of work.

Co-authored-by: Laveesh Rohra <larohra@microsoft.com>