Drop the interpreter for execution in the service file
The waagent script is executable and has the interpreter set. Drop the
use of the Python interpreter to execute the file from the systemd unit file.
We achieve unbuffered I/O by setting the PYTHONUNBUFFERED environment
variable instead. This makes it easier for distribution packaging as
packagers will only have to update the pathto the interpreter used on the
given distro in one file.