Swap out legacycrypt for crypt-r for Python 3.13+ (#3141)
PR #3070 pulled in legacycrypt to replace the removed crypt module.
legacycrypt hasn't been updated since it was initially pulled out of
Python's stdlib in 2019 (Python 3.7). crypt-r pulls in the module as it
was in Python 3.12. While there's been no major developments since 3.7,
it's more likely to be kept updated for any breakages in future Python
releases. It's also already packaged for Fedora and means one less
package for me to maintain so that would be nice.

As crypt-r provides a stub `crypt` module, the imports don't need to be
adjusted and WALinuxAgent can continue to unconditionally import from
`crypt`.

Co-authored-by: Norberto Arrieta <narrieta@users.noreply.github.com>